:recycle: refactor: best practice changes
-Refactored the fetching function to get 3 randomized pokemon, and
display their names as inputs. Use useCallback to wrap the newFetch
function, and newFetch is in the dependency for useEffect.
-Instead of extending a React component class, LandingPage is now a
function
-Capitalizing React components, and all of the corresponding imports.
-Delete out of scope files.
-Clean up the BattleStart code and get rid of unnecessary comments
-Add a form to ensure the start link only works if a pokemon has been
selected. If not, the start link should be disabled.
-The stray CSS folder for the initial "/" was causing build
failures when deploying to Netlify. Once deleted, the builds are
successful

PKM-3
