+++ package.json
@@ -17,8 +17,7 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject",
-    "compile:sass": "sass src/scss/base.scss src/css/styles.css"
+    "eject": "react-scripts eject"
   },
   "eslintConfig": {
     "extends": [
@@ -40,6 +39,6 @@
   },
   "devDependencies": {
     "cypress": "^12.0.1",
-    "sass": "^1.55.0"
+    "sass": "^1.57.1"
   }
 }
