+++ src/components/BattlePage.js
