+++ src/components/fetchPokemon.js
@@ -1,8 +1,8 @@
-// pokemonId = Math.round(Math.random() * 151);
+// let pokemonId = Math.round(Math.random() * 151);
 const fetchPokemon = async (pokemonId) => {
   const pokemonResponse = await fetch(`https://pokeapi.co/api/v2/pokemon/${pokemonId}`);
   const pokemonData = await pokemonResponse.json();
-  console.log(pokemonData);
+  // console.log(pokemonData);
   return pokemonData;
 }
