--- src/components/getAllPokemon.js
@@ -1,9 +1,0 @@
-import fetchPokemon from "./fetchPokemon";
-
-const getAllPokemon = () => {
-  for(let i = 1; i < 152; i++) {
-    fetchPokemon(i);
-  }
-}
-
-export default getAllPokemon;
