--- src/components/landingPage.js
@@ -1,72 +1,0 @@
-import React from "react";
-import { Link } from 'react-router-dom';
-
-class LandingPage extends React.Component {
-  render() {
-    return (
-      <>
-        <header className='page__header'>
-          <div className="header__title">
-            <h1>Welcome to </h1>
-            <svg width="314" height="44" fill="none" xmlns="http://www.w3.org/2000/svg">
-            <path d="M.479 7.517c1.496-.495 2.9-.953 4.211-1.373a74.032 74.032 0 0 1 4.175-1.188 46.861 46.861 0 0 1 4.194-.91 18.933 18.933 0 0 1 3.154-.259c1.793 0 2.969.05 3.525.149.841.16 1.633.42 2.375.779a6.803 6.803 0 0 1 2.004 1.447 7.132 7.132 0 0 1 1.447 2.115c.149.334.223.687.223 1.058 0 1.385-.76 3.092-2.282 5.121-1.51 2.029-3.03 3.798-4.565 5.307-1.521 1.509-2.66 2.591-3.414 3.247l3.34 11.708-4.861 3.507-8.517-24.326c-.024 0-.142.118-.352.353s-.489.544-.835.928c-.334.37-.544.6-.631.686-.037 0-.16-.266-.371-.798-.198-.531-.47-1.268-.817-2.208-.334-.952-.71-2.004-1.131-3.154C.918 8.543.627 7.813.479 7.516Zm12.004 5.158 1.522 5.473c.767-.779 1.72-1.843 2.857-3.191 1.138-1.348 1.707-2.325 1.707-2.932 0-.235-.099-.37-.296-.408a7.595 7.595 0 0 0-1.652-.185c-1.633 0-3.012.414-4.138 1.243Zm29.899 4.23a5.888 5.888 0 0 1 2.134 1.911 7.938 7.938 0 0 1 1.206 2.69c.26.99.39 2.023.39 3.1 0 .89-.075 1.75-.224 2.579-.136.828-.327 1.59-.575 2.282-.247.693-.538 1.28-.872 1.763-.779 1.162-1.614 2.164-2.505 3.005-.89.842-1.8 1.522-2.727 2.041-.928.507-1.85.879-2.765 1.114-.903.247-1.8.37-2.69.37a9.017 9.017 0 0 1-3.006-.5c-.94-.322-1.744-.749-2.412-1.28-1.101-.916-1.96-2.147-2.58-3.693a13.258 13.258 0 0 1-.927-4.935c0-2.982.86-5.387 2.579-7.218 1.472-1.571 2.993-2.753 4.564-3.544 1.572-.804 3.18-1.206 4.825-1.206.965 0 1.892.123 2.783.37.89.236 1.825.62 2.802 1.151Zm-1.095 3.117c-.21.94-.668 1.77-1.373 2.487-.693.717-1.373 1.076-2.041 1.076-.334 0-.631-.087-.89-.26-.372-.297-.557-.816-.557-1.558 0-.57.185-1.132.556-1.689.384-.556.984-.996 1.8-1.317-.816 0-1.608.154-2.375.464a6.566 6.566 0 0 0-1.948 1.206 6.015 6.015 0 0 0-1.318 1.651c-.321.606-.482 1.218-.482 1.837 0 1.15.532 2.066 1.596 2.746.99.63 1.93.946 2.82.946.878 0 1.658-.29 2.338-.872.68-.594 1.2-1.348 1.559-2.263.358-.928.538-1.843.538-2.746 0-.619-.075-1.188-.223-1.707Zm4.404-14.027 9.426-3.952 1.373 21.43 7.756-4.248 8.053 1.206-10.502 8.052 9.89 7.46.445 7.44-14.585-10.484.464 9.11-8.832.168-1.67-32.23-2.727.613.91-4.565ZM87.873 26.74l4.471 5.4c-.358 1.039-1.212 1.991-2.56 2.857-1.336.866-2.932 1.546-4.787 2.041-1.844.507-3.687.76-5.53.76-1.632 0-3.092-.216-4.379-.649-1.274-.433-2.257-1.088-2.95-1.966-1.2-1.51-1.8-3.44-1.8-5.79 0-1.855.334-3.68 1.002-5.473.68-1.806 1.398-3.068 2.153-3.785A13.868 13.868 0 0 1 76.424 18a14.374 14.374 0 0 1 3.303-1.28 14.404 14.404 0 0 1 3.507-.427c1.57 0 3.061.254 4.472.76 1.422.495 2.665 1.22 3.73 2.172L81.248 31.378c.903 0 1.756-.186 2.56-.557a7.771 7.771 0 0 0 2.245-1.614 12.133 12.133 0 0 0 1.819-2.468Zm-5.14-4.323a4.04 4.04 0 0 0-1.856-.464c-.667 0-1.317.136-1.948.408-.618.273-1.057.594-1.317.965-.767.99-1.15 2.078-1.15 3.266 0 .433.043.847.13 1.243.086.396.234.94.445 1.633l5.696-7.051ZM73.196 9.038l9.89-5.474 4.249 6.532-12.005 2.282-2.134-3.34Zm36.021 6.68c3.204-1.014 5.585-1.522 7.144-1.522 1.237 0 2.3.254 3.191.761 1.101.63 2.289 1.707 3.563 3.229 1.274 1.521 2.065 2.931 2.375 4.23.655 2.635.983 4.936.983 6.902 0 1.04-.068 2.134-.204 3.284a119.429 119.429 0 0 1-.631 4.398l-7.756.464c.025-.619.037-1.51.037-2.672 0-4.849-.668-8.974-2.004-12.376-.247-.606-.538-1.082-.872-1.429-.334-.346-.631-.52-.89-.52-.272 0-.409.193-.409.576 0 .223.025.414.075.575.086.198.173.464.259.798.668 2.684 1.002 4.985 1.002 6.902 0 .94-.061 1.93-.185 2.97-.124 1.026-.34 2.597-.65 4.712l-7.755.464c.024-.619.037-1.51.037-2.672 0-4.861-.674-8.987-2.023-12.376a6.896 6.896 0 0 1-.371-1.243c-.086-.445-.148-.668-.185-.668-.05 0-.192.29-.427.872a64.948 64.948 0 0 1-.761 1.8c-.185.433-.34.928-.464 1.484a17.714 17.714 0 0 0-.26 1.874c-.049.693-.074 1.51-.074 2.45 0 1.904.112 4.323.334 7.254l-9.11 1.058c-.149-1.089-.285-2.171-.408-3.247a89.918 89.918 0 0 1-.297-2.987 45.078 45.078 0 0 1-.093-2.988c0-.457.025-1.12.074-1.985l-.334-10.817 7.292-.928 1.058 1.224c1.954-.63 3.55-.946 4.787-.946 1.373 0 2.69.365 3.952 1.095Zm35.02 1.187a5.885 5.885 0 0 1 2.133 1.911 7.922 7.922 0 0 1 1.206 2.69c.26.99.39 2.023.39 3.1 0 .89-.074 1.75-.223 2.579a13.03 13.03 0 0 1-.575 2.282c-.247.693-.538 1.28-.872 1.763-.779 1.162-1.614 2.164-2.505 3.005-.89.842-1.8 1.522-2.727 2.041a12 12 0 0 1-2.765 1.114c-.903.247-1.8.37-2.69.37a9.014 9.014 0 0 1-3.006-.5c-.94-.322-1.744-.749-2.412-1.28-1.101-.916-1.961-2.147-2.579-3.693a13.248 13.248 0 0 1-.928-4.935c0-2.982.859-5.387 2.579-7.218 1.472-1.571 2.993-2.753 4.564-3.544 1.571-.804 3.179-1.206 4.825-1.206.964 0 1.892.123 2.783.37.89.236 1.824.62 2.802 1.151Zm-1.095 3.117c-.21.94-.668 1.77-1.373 2.487-.693.717-1.373 1.076-2.041 1.076a1.57 1.57 0 0 1-.891-.26c-.371-.297-.557-.816-.557-1.558 0-.57.186-1.132.557-1.689.384-.556.984-.996 1.8-1.317-.816 0-1.608.154-2.375.464a6.556 6.556 0 0 0-1.948 1.206 6.003 6.003 0 0 0-1.318 1.651c-.321.606-.482 1.218-.482 1.837 0 1.15.532 2.066 1.596 2.746.989.63 1.929.946 2.82.946.878 0 1.657-.29 2.338-.872.68-.594 1.2-1.348 1.558-2.263.359-.928.538-1.843.538-2.746 0-.619-.074-1.188-.222-1.707Zm13.978-4.453c1.373-.457 2.555-.804 3.544-1.039.99-.247 1.936-.37 2.839-.37 1.212 0 2.276.265 3.191.797 1.101.63 2.289 1.707 3.563 3.229 1.274 1.521 2.066 2.931 2.375 4.23.655 2.635.983 4.936.983 6.902 0 1.04-.068 2.134-.204 3.284A119.429 119.429 0 0 1 172.78 37l-7.756.464c.025-.619.038-1.51.038-2.672 0-4.849-.668-8.974-2.004-12.376-.52-1.274-1.027-1.911-1.522-1.911-.643 0-1.354.89-2.134 2.672a8.39 8.39 0 0 0-.463 1.484 17.457 17.457 0 0 0-.26 1.874 34.65 34.65 0 0 0-.074 2.45c0 1.904.111 4.323.334 7.254l-9.111 1.058c-.148-1.089-.284-2.171-.408-3.247a88.612 88.612 0 0 1-.297-2.987 44.936 44.936 0 0 1-.093-2.988c0-.457.025-1.12.075-1.985l-.334-10.817 7.292-.928 1.057 1.224Zm64.695-1.985-8.312 5.436c-.557-1.385-2.419-2.078-5.585-2.078-.718 0-1.454.514-2.208 1.54a5.711 5.711 0 0 0-1.132 3.47c.062.977.352 1.917.872 2.82.519.891 1.441 1.367 2.765 1.43.989 0 2.028-.267 3.117-.799 1.088-.532 2.028-1.595 2.82-3.191.05.099.563.563 1.54 1.392.977.828 1.973 1.657 2.987 2.486 1.027.829 1.782 1.44 2.264 1.837-.94 2.251-2.183 4.014-3.729 5.288-1.547 1.274-3.167 2.146-4.862 2.616-1.694.47-3.377.705-5.047.705-2.338 0-4.397-.439-6.178-1.317-1.769-.89-3.223-2.406-4.361-4.546-1.138-2.152-1.83-5.06-2.078-8.72a25.854 25.854 0 0 1-.037-1.634c0-1.83.217-3.587.65-5.27a13.242 13.242 0 0 1 2.078-4.545c.977-1.36 2.319-2.468 4.026-3.321 1.707-.854 3.773-1.3 6.197-1.336 1.608 0 3.192.191 4.75.575 1.559.371 2.988.897 4.286 1.577 1.299.668 2.413 1.49 3.34 2.468.94.965 1.553 2.004 1.837 3.117Zm10.212 28.129-8.368.464-1.67-32.23-2.728.613.909-4.565 9.426-3.952 2.431 39.67Zm14.145-22.878a6.03 6.03 0 0 0-2.022.65 6.962 6.962 0 0 0-1.689 1.224c-.47.47-.841.99-1.113 1.559a3.879 3.879 0 0 0-.39 1.65c0 1.127.532 2.042 1.596 2.747.94.63 1.893.946 2.858.946.185 0 .439-.024.76-.074v-3.952c-.334 0-.631-.087-.89-.26-.371-.297-.557-.866-.557-1.707 0-1.212.482-2.14 1.447-2.783Zm.149-3.266 7.756-3.488.148 10.372c.099.643.148 1.373.148 2.19 0 .717-.024 1.249-.074 1.595l.223 11.374-8.35.446v-1.132c-1.36.556-2.733.835-4.119.835a9.223 9.223 0 0 1-3.006-.483c-.927-.321-1.731-.754-2.412-1.298-1.101-.916-1.96-2.147-2.579-3.693a13.268 13.268 0 0 1-.928-4.935c0-2.982.86-5.387 2.579-7.218 1.485-1.559 3.012-2.74 4.583-3.544 1.584-.804 3.198-1.206 4.843-1.206.458 0 .804.012 1.039.037l.149.148Zm30.009 7.292c-.371-.136-1.274-.42-2.709-.853a159.684 159.684 0 0 0-3.822-1.15 59.373 59.373 0 0 0-3.692-.928c-1.349-.297-2.4-.446-3.155-.446-.395 0-.699.062-.909.186-.606.458-.909.878-.909 1.262 0 .606.606 1.132 1.818 1.577 1.213.433 2.685.754 4.416.965 2.375.26 4.293.816 5.752 1.67 1.472.853 2.536 1.849 3.192 2.987.655 1.125.983 2.282.983 3.47 0 .84-.167 1.626-.501 2.356-1.435 3.043-4.187 4.565-8.257 4.565-1.113 0-2.368-.124-3.766-.372a39.189 39.189 0 0 1-4.027-.89 88.426 88.426 0 0 1-4.082-1.225c-1.447-.47-2.795-.909-4.045-1.317.322-.668.73-1.472 1.225-2.412l1.447-2.783c.458-.904.687-1.398.687-1.485.359.334.804.606 1.336.817.544.197 1.33.439 2.356.723 1.039.272 1.831.52 2.375.742 1.2.446 2.443.668 3.73.668.791 0 1.546-.08 2.263-.24.718-.174 1.101-.316 1.151-.428 0-.321-.551-.637-1.652-.946-1.088-.322-2.307-.569-3.655-.742a20.074 20.074 0 0 1-4.639-1.113c-1.546-.582-2.863-1.43-3.952-2.542-1.088-1.114-1.633-2.487-1.633-4.12 0-1.249.316-2.554.947-3.915.692-1.521 1.732-2.622 3.117-3.302 1.385-.693 3.043-1.04 4.973-1.04 2.226 0 4.521.316 6.883.947 2.363.63 5.579 1.608 9.649 2.931-.087.186-.427.978-1.021 2.375-.593 1.386-1.039 2.4-1.336 3.043-.296.644-.476.965-.538.965Zm11.103-7.292c2.214-.73 4.125-1.094 5.733-1.094 1.472 0 2.752.315 3.841.946.656.383 1.392.965 2.208 1.744a14.551 14.551 0 0 1 2.152 2.542c.631.928 1.052 1.83 1.262 2.709.656 2.635.983 4.936.983 6.902 0 1.04-.068 2.134-.204 3.284a119.429 119.429 0 0 1-.631 4.398l-7.755.464c.024-.619.037-1.51.037-2.672 0-4.849-.668-8.974-2.004-12.376-.52-1.274-1.027-1.911-1.522-1.911-.643 0-1.354.89-2.133 2.672a8.307 8.307 0 0 0-.464 1.484 17.457 17.457 0 0 0-.26 1.874 34.65 34.65 0 0 0-.074 2.45c0 1.904.111 4.323.334 7.254l-9.111 1.058a22.875 22.875 0 0 1-.185-2.524c-.037-.964-.062-1.96-.074-2.987a58.706 58.706 0 0 0-.13-2.802 29.19 29.19 0 0 0-.334-2.894l-.501-24.195 7.607-.928 1.225 14.602Zm20.157 13.675c.878 0 1.633.149 2.263.445.644.285 1.169.637 1.578 1.058.408.408.686.767.835 1.076.148.347.222.718.222 1.114a4.73 4.73 0 0 1-.389 1.836 6.752 6.752 0 0 1-.965 1.615c-.371.457-.718.711-1.039.76-.668.1-1.151.149-1.448.149-1.224 0-2.208-.316-2.95-.946-.73-.644-1.163-1.516-1.299-2.617a6.03 6.03 0 0 1 .761-2.968c.52-.916 1.33-1.423 2.431-1.522Zm-2.746-5.158-2.431-20.522 11.096-3.34-2.283 23.25-6.382.612Z" fill="#FFDB38"/>
-            </svg>
-            <h2>super simple Pokémon battles</h2>
-          </div>
-        
-          <nav className="header-nav">
-            <ul className="header-nav__list">
-              <li>
-                <Link to="/pokemonList" aria-current="page">
-                  See All Pokémon
-                </Link>
-              </li>
-            </ul>
-          </nav>
-        </header>
-        
-        <article className="rules" id="battle-rules">
-          <header className='rules__header'>
-            <h3>Rules</h3>
-          </header>
-          {/* <!-- selection rules --> */}
-          <div className="rules-flex-container">
-            <section className="rules__before-battle">
-              <h4>
-                Before the battle...
-              </h4>
-              <ul className="rules__card rules-1 ">
-                <li>Pick your Pokémon</li>
-                <li>See your opponent</li>
-              </ul>
-            </section>
-            {/* <!-- battle rules (picking moves, attacking, dying) --> */}
-            <section className="rules__during-battle">
-              <h4>
-                During the battle...
-              </h4>
-              <ul className="rules__card rules-2">
-                <li>Select your moves</li>
-                <li>Attack your opponent</li>
-                <li><span className="uppercase">Beware:</span> you will be attacked!</li>
-              </ul>
-            </section>
-            {/* <!-- how to win --> */}
-            <section className="rules__end-battle">
-              <h4>Don't die first!</h4>
-              <ul class="rules__card rules-3">
-                <li>Stay alive longer than your opponent to win!</li>
-              </ul>
-            </section>
-          </div>
-
-          <Link className="start-button" to="/battleStart">
-              Let's Battle!
-          </Link>
-        </article>
-      </>
-    )
-  }
-}
-
-export default LandingPage;
