+++ src/scss/base.scss
@@ -28,6 +28,7 @@
 @import "./components/landingPage.scss";
 @import "./components/startButton.scss";
 @import "./components/battleStart";
+@import "./components/app.scss";
 
 /*VENDORS*/
 /*third-party*/
