+++ src/scss/components/app.scss
@@ -1,0 +1,198 @@
+/*SETTINGS*/
+/* variables, custom properties */
+/*TOOLS*/
+/*mixins*/
+/*GENERIC*/
+/* fonts, reset stylesheets */
+/*ELEMENTS*/
+/* html element selectors only */
+html {
+  color: #ffdb38;
+  background-color: #1a226b;
+}
+
+body {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  padding-top: 3rem;
+  box-sizing: border-box;
+  font-family: "Source Sans Pro", sans-serif;
+}
+
+/*LAYOUT/OBJECTS*/
+/*objects, grids, flexbox etc*/
+/*COMPONENTS*/
+/*most of the work goes here! BEM classes, headers/footers/carousels, etc*/
+.page__header {
+  text-align: center;
+  display: flex;
+  flex-direction: column;
+}
+.page__header h1 {
+  font-size: 3rem;
+  padding-bottom: 1.5rem;
+}
+.page__header h2 {
+  font-size: 1.5rem;
+  padding-top: 2.5rem;
+}
+@media only screen and (min-width: 550px) {
+  .page__header svg {
+    scale: 150%;
+  }
+}
+@media only screen and (min-width: 850px) {
+  .page__header svg {
+    scale: 200%;
+  }
+}
+@media only screen and (min-width: 1140px) {
+  .page__header svg {
+    scale: 250%;
+  }
+}
+
+ul {
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
+  margin-top: 1rem;
+}
+ul a {
+  text-decoration: none;
+  transition: all 100ms ease-in-out;
+  font-weight: 600;
+  font-size: 1.25rem;
+  color: #ffdb38;
+}
+ul a:hover {
+  color: #fefefe;
+}
+
+.rules {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  position: relative;
+  top: 1.5rem;
+}
+@media only screen and (min-width: 550px) {
+  .rules {
+    margin-top: 2rem;
+  }
+}
+.rules-flex-container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  flex-basis: 100%;
+}
+@media only screen and (min-width: 680px) {
+  .rules-flex-container {
+    flex-direction: row;
+    justify-content: space-around;
+    align-items: flex-start;
+  }
+}
+.rules__before-battle, .rules__during-battle, .rules__end-battle {
+  flex-basis: 33%;
+  padding: 0 0.75rem 0.75rem 0.75rem;
+  text-align: left;
+}
+.rules h4 {
+  font-size: 1.375rem;
+  margin-top: 0.5rem;
+  margin-bottom: 0;
+}
+.rules .rules-1,
+.rules .rules-2,
+.rules .rules-3 {
+  margin-top: 0.5rem;
+  padding-bottom: 1rem;
+}
+
+.uppercase {
+  text-transform: uppercase;
+}
+
+.rules__card {
+  list-style-image: url("../../resources/pokeball_list-image.svg");
+}
+.rules__card li {
+  padding-left: 0.25rem;
+}
+@media only screen and (min-width: 680px) {
+  .rules__card li {
+    padding: 0 0 0 0.5rem;
+  }
+}
+
+.start-button {
+  background-color: transparent;
+  color: #ffdb38;
+  font-size: 1.625rem;
+  border: 2px solid #ffdb38;
+  padding: 0.75em;
+  border-radius: 3rem;
+  transition: all 100ms ease-in-out;
+  position: relative;
+  top: 4rem;
+  cursor: pointer;
+  text-decoration: none;
+}
+.start-button:hover {
+  color: white;
+  box-shadow: 4px 0px 4px 4px rgba(0, 0, 0, 0.1411764706);
+  background-color: #EE3918;
+  border: 2px solid white;
+}
+
+.start-button {
+  background-color: transparent;
+  color: #ffdb38;
+  font-size: 1.625rem;
+  border: 2px solid #ffdb38;
+  padding: 0.75em;
+  border-radius: 3rem;
+  transition: all 100ms ease-in-out;
+  position: relative;
+  top: 4rem;
+  cursor: pointer;
+  text-decoration: none;
+}
+.start-button:hover {
+  color: white;
+  box-shadow: 4px 0px 4px 4px rgba(0, 0, 0, 0.1411764706);
+  background-color: #EE3918;
+  border: 2px solid white;
+}
+
+.enter-battle {
+  display: inline-block;
+  margin: auto;
+}
+
+/*VENDORS*/
+/*third-party*/
+/*UTILITIES*/
+/*one off, hiddens, backgrounds, margin, padding, etc*/
+.enter {
+  border-bottom: 0.25rem solid #EE3918;
+  font-family: "Source Sans Pro", sans-serif;
+  font-weight: 400;
+  font-size: 3rem;
+  color: #ffdb38;
+  transition: all 100ms ease-in-out;
+  text-decoration: none;
+}
+.enter:hover {
+  color: #fff;
+  font-weight: 700;
+  border-bottom: 0.375rem solid #fff;
+}
+
+/*# sourceMappingURL=styles.css.map */
